Update spindle speed cvontrol for higher resolution
You can now change the spindle PWM resolution.
Basic instructions have been added to cpu_map.h
Default frequency is still 5000 Hz
Resolution has been changed fro 8 bit to 12 bit.

+++ Grbl_Esp32/spindle_control.cpp
@@ -36,7 +36,7 @@ void spindle_init()
 	#endif
 	
 	#ifdef SPINDLE_PWM_PIN
-    // use the LED control feature to setup PWM   https://esp-idf.readthedocs.io/en/v1.0/api/ledc.html
+    // use the LED control feature to setup PWM   https://docs.espressif.com/projects/esp-idf/en/latest/api-reference/peripherals/ledc.html
     ledcSetup(SPINDLE_PWM_CHANNEL, SPINDLE_PWM_BASE_FREQ, SPINDLE_PWM_BIT_PRECISION); // setup the channel
     ledcAttachPin(SPINDLE_PWM_PIN, SPINDLE_PWM_CHANNEL); // attach the PWM to the pin
 	#endif
@@ -75,7 +75,7 @@ uint8_t spindle_get_state()  // returns SPINDLE_STATE_DISABLE, SPINDLE_STATE_CW 
 	}
 }
 
-void spindle_set_speed(uint8_t pwm_value)
+void spindle_set_speed(uint32_t pwm_value)
 {	
 	#ifndef SPINDLE_PWM_PIN
 		return;
@@ -90,9 +90,9 @@ void spindle_set_speed(uint8_t pwm_value)
 }
 
 // Called by spindle_set_state() and step segment generator. Keep routine small and efficient.
-uint8_t spindle_compute_pwm_value(float rpm)
+uint32_t spindle_compute_pwm_value(float rpm)
 {
-  uint8_t pwm_value;
+  uint32_t pwm_value;
   rpm *= (0.010*sys.spindle_speed_ovr); // Scale by spindle speed override value.
   // Calculate PWM register value based on rpm max/min settings and programmed rpm.
   if ((settings.rpm_min >= settings.rpm_max) || (rpm >= settings.rpm_max)) {
@@ -114,8 +114,6 @@ uint8_t spindle_compute_pwm_value(float rpm)
     pwm_value = floor((rpm-settings.rpm_min)*pwm_gradient) + SPINDLE_PWM_MIN_VALUE;
   }
   return(pwm_value);
-
-
 
 }
 
@@ -156,6 +154,7 @@ void grbl_analogWrite(uint8_t chan, uint32_t duty)
 {
 	if (ledcRead(chan) != duty) // reduce unnecessary calls to ledcWrite()
 	{
+		// grbl_sendf(CLIENT_SERIAL, "[MSG: Spindle duty: %d of %d]\r\n", duty, SPINDLE_PWM_MAX_VALUE); // debug statement
 		ledcWrite(chan, duty);
 	}
 }
